+++ swarm/src/lib.rs
@@ -359,20 +359,22 @@ where
 
     fn dial_with_handler(
         &mut self,
-        opts: DialOpts,
+        swarm_dial_opts: DialOpts,
         handler: <TBehaviour as NetworkBehaviour>::ProtocolsHandler,
     ) -> Result<(), DialError> {
-        match opts.0 {
+        let core_dial_opts = match swarm_dial_opts.0 {
             // Dial a known peer.
             dial_opts::Opts::WithPeerId(dial_opts::WithPeerId {
                 peer_id,
                 condition,
                 role_override,
+                dial_concurrency_factor_override,
             })
             | dial_opts::Opts::WithPeerIdWithAddresses(dial_opts::WithPeerIdWithAddresses {
                 peer_id,
                 condition,
                 role_override,
+                dial_concurrency_factor_override,
                 ..
             }) => {
                 // Check [`PeerCondition`] if provided.
@@ -401,7 +403,7 @@ where
 
                 // Retrieve the addresses to dial.
                 let addresses = {
-                    let mut addresses = match opts.0 {
+                    let mut addresses = match swarm_dial_opts.0 {
                         dial_opts::Opts::WithPeerId(dial_opts::WithPeerId { .. }) => {
                             self.behaviour.addresses_of_peer(&peer_id)
                         }
@@ -438,56 +440,37 @@ where
                     addresses
                 };
 
-                let handler = handler
-                    .into_node_handler_builder()
-                    .with_substream_upgrade_protocol_override(
-                        self.substream_upgrade_protocol_override,
-                    );
+                let mut opts = libp2p_core::DialOpts::peer_id(peer_id)
+                    .addresses(addresses)
+                    .override_role(role_override);
 
-                match self
-                    .network
-                    .peer(peer_id)
-                    .dial(addresses, handler, role_override)
-                {
-                    Ok(_connection_id) => Ok(()),
-                    Err(error) => {
-                        let (error, handler) = DialError::from_network_dial_error(error);
-                        self.behaviour.inject_dial_failure(
-                            Some(peer_id),
-                            handler.into_protocols_handler(),
-                            &error,
-                        );
-                        return Err(error);
-                    }
+                if let Some(f) = dial_concurrency_factor_override {
+                    opts = opts.override_dial_concurrency_factor(f);
                 }
+
+                opts.build()
             }
             // Dial an unknown peer.
             dial_opts::Opts::WithoutPeerIdWithAddress(dial_opts::WithoutPeerIdWithAddress {
                 address,
                 role_override,
-            }) => {
-                let handler = handler
-                    .into_node_handler_builder()
-                    .with_substream_upgrade_protocol_override(
-                        self.substream_upgrade_protocol_override,
-                    );
+            }) => libp2p_core::DialOpts::unknown_peer_id()
+                .address(address)
+                .override_role(role_override)
+                .build(),
+        };
 
-                match self
-                    .network
-                    .dial(&address, handler, role_override)
-                    .map(|_id| ())
-                {
-                    Ok(_connection_id) => Ok(()),
-                    Err(error) => {
-                        let (error, handler) = DialError::from_network_dial_error(error);
-                        self.behaviour.inject_dial_failure(
-                            None,
-                            handler.into_protocols_handler(),
-                            &error,
-                        );
-                        return Err(error);
-                    }
-                }
+        let handler = handler
+            .into_node_handler_builder()
+            .with_substream_upgrade_protocol_override(self.substream_upgrade_protocol_override);
+
+        match self.network.dial(handler, core_dial_opts).map(|_id| ()) {
+            Ok(_connection_id) => Ok(()),
+            Err(error) => {
+                let (error, handler) = DialError::from_network_dial_error(error);
+                self.behaviour
+                    .inject_dial_failure(None, handler.into_protocols_handler(), &error);
+                return Err(error);
             }
         }
     }
@@ -719,9 +702,18 @@ where
                     handler,
                 }) => {
                     if let Some(error) = error.as_ref() {
-                        log::debug!("Connection {:?} closed: {:?}", connected, error);
+                        log::debug!(
+                            "Connection closed with error {:?}: {:?}; Total (peer): {}.",
+                            error,
+                            connected,
+                            remaining_established_connection_ids.len()
+                        );
                     } else {
-                        log::debug!("Connection {:?} closed (active close).", connected);
+                        log::debug!(
+                            "Connection closed: {:?}; Total (peer): {}.",
+                            connected,
+                            remaining_established_connection_ids.len()
+                        );
                     }
                     let peer_id = connected.peer_id;
                     let endpoint = connected.endpoint;
@@ -1339,8 +1331,9 @@ pub enum DialError {
     DialPeerConditionFalse(dial_opts::PeerCondition),
     /// Pending connection attempt has been aborted.
     Aborted,
-    /// The peer identity obtained on the connection did not
-    /// match the one that was expected or is otherwise invalid.
+    /// The provided peer identity is invalid or the peer identity obtained on
+    /// the connection did not match the one that was expected or is otherwise
+    /// invalid.
     InvalidPeerId,
     /// An I/O error occurred on the connection.
     ConnectionIo(io::Error),
@@ -1355,6 +1348,7 @@ impl DialError {
                 (DialError::ConnectionLimit(limit), handler)
             }
             network::DialError::LocalPeerId { handler } => (DialError::LocalPeerId, handler),
+            network::DialError::InvalidPeerId { handler } => (DialError::InvalidPeerId, handler),
         }
     }
 }