+++ protocols/kad/CHANGELOG.md
@@ -4,7 +4,13 @@
 
 - Migrate to Rust edition 2021 (see [PR 2339]).
 
+- Derive `Clone` for `KademliaEvent` (see [PR 2411])
+
+- Derive `Serialize`, `Deserialize` for `store::record::Key` (see [PR 2408])
+
 [PR 2339]: https://github.com/libp2p/rust-libp2p/pull/2339
+[PR 2411]: https://github.com/libp2p/rust-libp2p/pull/2411
+[PR 2408]: https://github.com/libp2p/rust-libp2p/pull/2408
 
 # 0.33.0 [2021-11-16]